refactor: separate media matching pipeline
